Fix server-side API route crashes by separating React hook from utility functions
- Moved useAuth import inside useDataAccess hook to prevent server-side crashes
- Separated React hook (useDataAccess) into dedicated hooks file
- Kept utility functions in data-access.ts for server-side API routes
- Server-side API routes can now safely import utility functions without React hook dependencies
- Added proper error handling for server-side rendering scenarios
